% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni-GEN.R
\name{classifications}
\alias{classifications}
\title{Get geographical classifications of municipalities}
\usage{
classifications(
  start_date = NULL,
  end_date = NULL,
  historicized_code = FALSE,
  name_type = c("ID", "en", "de", "fr", "it"),
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{start_date}{The begin of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{end_date}{The end of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{historicized_code}{By default, the \href{https://de.wikipedia.org/wiki/Gemeindenummer}{\emph{FSO commune number}} is returned. Set to \code{TRUE} in order to get the
\emph{historicization number} instead.}

\item{name_type}{The column name type to be returned. Defaults to the column IDs returned by the web service. If set to a language code, the columns
starting with ID \code{HR_} will be named by its full label in the respective language instead.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
This function returns a \link[tibble:tbl_df]{tibble} with a wide range of geographical classifications of Swiss municipalities.
}
\details{
There are some classifications which have multiple column IDs (e.g. \code{HR_ARBREG2000_L2} and \code{HR_TYPMSR2000_L2} both contain the \emph{spatial mobility regions}).
If \code{name_type} is set to a language code, column duplicates are given the name suffix " (⎘)".

\subsection{FSO commune number vs. historicization number}{

By default, the \href{https://de.wikipedia.org/wiki/Gemeindenummer}{\emph{FSO commune number}} is returned. You
can set \code{historicized_code = TRUE} to get the \emph{historicization number}
instead. Details about the differences between the both can be found in
\href{https://www.bfs.admin.ch/bfs/de/home/grundlagen/agvch/historisiertes-gemeindeverzeichnis.assetdetail.4062821.html}{this PDF}
(in German). A translation of the relevant paragraph (p. 2):

―

The mutation processes at the cantonal, district and municipal level
cause changes in the official register of Swiss municipalities. The
so-called historicization number uniquely identifies the entries in the
\href{https://www.bfs.admin.ch/bfs/en/home/basics/swiss-official-commune-register.html}{official commune register}
that have been changed, “deleted” or added as a result of mutation
processes.

In contrast to the \strong{FSO commune number} or district number, the
\strong{historicization number} is \emph{unique within the database} and defines
the state of the communes or districts during a certain period of time.
The mutation numbers, the type of inclusion or revocation and the date
of inclusion or revocation are used to fully describe the individual
mutation events in the “Municipalities” and “Districts” tables. The
mutation number identifies the individual mutation events that may
affect one or more entries.

The historicization number does not replace the existing municipality
and district numbers. It is used for the computer-aided mapping of the
various states and mutation processes of the official commune register
and the districts.

―
}
}
\examples{
classifications(start_date = lubridate::today(),
                end_date = lubridate::today(),
                name_type = "de",
                use_cache = FALSE)
}
