% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni.gen.R
\name{congruences}
\alias{congruences}
\title{Get municipality congruences}
\usage{
congruences(
  start_date = NULL,
  end_date = NULL,
  incl_unmodified = TRUE,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)
}
\arguments{
\item{start_date}{Begin of the period the data covers. A \link[base:Dates]{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{end_date}{End of the period the data covers. A \link[base:Dates]{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{incl_unmodified}{Include municipalities where no changes occurred in the defined period.}

\item{incl_territory_exchange}{Include information (additional rows) about territory exchanges within and between municipalities.}

\item{use_cache}{Whether or not to return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{Duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with congruence data of Swiss municipalities. The data indicates for each existing municipality at the \code{start_date} to
which municipality/municipalities it corresponds at the \code{end_date}.
}
\examples{
swissmuni::congruences(start_date = lubridate::today(),
                       end_date = lubridate::today(),
                       max_cache_age = "6 hours")
}
