% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni.gen.R
\name{snapshots}
\alias{snapshots}
\title{Get municipality snapshots}
\usage{
snapshots(
  start_date = lubridate::today(),
  end_date = start_date,
  historicized_code = FALSE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)
}
\arguments{
\item{start_date}{The begin of the period the data covers. A \link[base:Dates]{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{end_date}{The end of the period the data covers. A \link[base:Dates]{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{historicized_code}{By default, the \href{https://de.wikipedia.org/wiki/Gemeindenummer}{\emph{FSO commune number}} is returned. Set to \code{TRUE} in order to get the
\emph{historicization number} instead.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with snapshots of Swiss municipality data. A snapshot corresponds to the list of municipalities that are present at least
partially during the specified period (or at the specified day when \code{end_date = start_date}).
}
\details{
\subsection{FSO commune number vs. historicization number}{

By default, the \href{https://de.wikipedia.org/wiki/Gemeindenummer}{\emph{FSO commune number}} is returned. You
can set \code{historicized_code = TRUE} to get the \emph{historicization number}
instead. Details about the differences between the both can be found in
\href{https://www.bfs.admin.ch/bfs/de/home/grundlagen/agvch/historisiertes-gemeindeverzeichnis.assetdetail.4062821.html}{this PDF}
(in German). A translation of the relevant paragraph (p. 2):

―

The mutation processes at the cantonal, district and municipal level
cause changes in the official register of Swiss municipalities. The
so-called historicization number uniquely identifies the entries in the
\href{https://www.bfs.admin.ch/bfs/en/home/basics/swiss-official-commune-register.html}{official commune register}
that have been changed, “deleted” or added as a result of mutation
processes.

In contrast to the \strong{FSO commune number} or district number, the
\strong{historicization number} is \emph{unique within the database} and defines
the state of the communes or districts during a certain period of time.
The mutation numbers, the type of inclusion or revocation and the date
of inclusion or revocation are used to fully describe the individual
mutation events in the “Municipalities” and “Districts” tables. The
mutation number identifies the individual mutation events that may
affect one or more entries.

The historicization number does not replace the existing municipality
and district numbers. It is used for the computer-aided mapping of the
various states and mutation processes of the official commune register
and the districts.

―
}
}
\examples{
swissmuni::snapshots(historicized_code = TRUE)
}
