% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni-package.R
\docType{package}
\name{swissmuni-package}
\alias{swissmuni}
\alias{swissmuni-package}
\title{swissmuni: Download Municipality Data from the Swiss Federal
    Statistical Office's Web Services}
\description{
swissmuni provides functions to access
    municipality snapshots, congruences, mutations and their geographical
    classifications from the web services provided by the Swiss Federal
    Statistical Office (FSO). These web services are part of the FSO's
    Statistical Metadata System (SMS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/swissmuni}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/swissmuni/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
