% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni-GEN.R
\name{api_params}
\alias{api_params}
\title{Get allowed API URL parameters}
\usage{
api_params(
  type = c("snapshots", "congruences", "mutations", "classifications")
)
}
\arguments{
\item{type}{The API endpoint.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Get allowed API URL parameters
}
\examples{
# to get all URL params of all 3 API endpoints:
library(magrittr)

endpoints <- c("snapshots",
               "congruences",
               "mutations",
               "classifications")

\donttest{
endpoints \%>\%
  purrr::map(getFromNamespace("api_params",
                              ns = "swissmuni")) \%>\%
  rlang::set_names(nm = endpoints)}
}
\keyword{internal}
