% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni-GEN.R
\name{congruences}
\alias{congruences}
\title{Get municipality congruences}
\usage{
congruences(
  start_date = NULL,
  end_date = NULL,
  incl_unmodified = TRUE,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{start_date}{The begin of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{end_date}{The end of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{incl_unmodified}{Include municipalities where no changes occurred in the defined period.}

\item{incl_territory_exchange}{Include information (additional rows) about territory exchanges within and between municipalities.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
This function returns a \link[tibble:tbl_df-class]{tibble} with congruence data of Swiss municipalities.
}
\examples{
congruences(start_date = lubridate::today(),
            end_date = lubridate::today(),
            cache_lifespan = "6 hours")
}
