% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmuni-GEN.R
\name{mutations}
\alias{mutations}
\title{Get municipality mutations}
\usage{
mutations(
  start_date = NULL,
  end_date = NULL,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)
}
\arguments{
\item{start_date}{The begin of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{end_date}{The end of the period the data covers. A \href{base::Date}{date} or a character scalar in the format \code{"YYYY-MM-DD"}.}

\item{incl_territory_exchange}{Include information about territory exchanges between municipalities.}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}. Defaults to 1 day (24 hours).}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
This function returns a \link[tibble:tbl_df]{tibble} with mutation data of Swiss municipalities.
}
\examples{
mutations(start_date = "2020-01-01",
          end_date = "2020-06-01")
}
